+++ lte/alt1250/usock_handlers/alt1250_ioctl_fwupdate.c
@@ -55,7 +55,7 @@ static int postproc_fwgetimglen(FAR struct alt1250_s *dev,
                                 FAR struct alt_container_s *reply,
                                 FAR struct usock_s *usock,
                                 FAR int32_t *usock_result,
-                                FAR uint8_t *usock_xid,
+                                FAR uint64_t *usock_xid,
                                 FAR struct usock_ackinfo_s *ackinfo,
                                 unsigned long arg)
 {
@@ -132,7 +132,7 @@ static int postproc_fwupdate_injection(FAR struct alt1250_s *dev,
                                        FAR struct alt_container_s *reply,
                                        FAR struct usock_s *usock,
                                        FAR int32_t *usock_result,
-                                       FAR uint8_t *usock_xid,
+                                       FAR uint64_t *usock_xid,
                                        FAR struct usock_ackinfo_s *ackinfo,
                                        unsigned long arg)
 {
@@ -161,7 +161,7 @@ static int postproc_fwupdate_dummy(FAR struct alt1250_s *dev,
                                   FAR struct alt_container_s *reply,
                                   FAR struct usock_s *usock,
                                   FAR int32_t *usock_result,
-                                  FAR uint8_t *usock_xid,
+                                  FAR uint64_t *usock_xid,
                                   FAR struct usock_ackinfo_s *ackinfo,
                                   unsigned long arg)
 {
@@ -321,7 +321,7 @@ static int fwupdate_header_injection(FAR struct alt1250_s *dev,
 int usockreq_ioctl_fwupdate(FAR struct alt1250_s *dev,
                             FAR struct usrsock_request_buff_s *req,
                             FAR int32_t *usock_result,
-                            FAR uint8_t *usock_xid,
+                            FAR uint64_t *usock_xid,
                             FAR struct usock_ackinfo_s *ackinfo)
 {
   FAR struct usrsock_request_ioctl_s *request = &req->request.ioctl_req;
